fixing more warnings, explicit casts, etc
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4874 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ common/include/pcl/common/impl/common.hpp
@@ -62,8 +62,8 @@ pcl::getMeanStd (const std::vector<float> &values, double &mean, double &stddev)
     sum += values[i];
     sq_sum += values[i] * values[i];
   }
-  mean = sum / values.size ();
-  double variance = (double)(sq_sum - sum * sum / values.size ()) / (values.size () - 1);
+  mean = sum / static_cast<double>(values.size ());
+  double variance = (sq_sum - sum * sum / static_cast<double>(values.size ())) / (static_cast<double>(values.size ()) - 1);
   stddev = sqrt (variance);
 }
